do not display language selector if lonly one language provided
